requirements.txt
python-telegram-bot==13.15
requests==2.31.0
beautifulsoup4==4.12.3